Improve the wiki command by using the text from the Intro section and taking the first text-looking line
Two things needed to be changed in order for the wiki command to work on the two presented error cases ("USWNT" and "penis").
The first issue was we were taking the first section returned in the results. The problem is that the first section wasn't always the Intro section. The sections were sorted alphabetically so it was often the case that Intro wasn't the first one. Now we explictly take the Intro section, which, as far as I can tell, always exists.
The second issue was ignoring lines that look like wikipedia formatting (e.g. start with a "{{" and "<"). There are probably more formatting characters but these should be a good start.